refactor: consolidate Rhino layer and object utilities into helpers
