+++ Connectors/Rhino/Speckle.Connectors.RhinoShared/Speckle.Connectors.RhinoShared.projitems
@@ -25,6 +25,8 @@
     <Compile Include="$(MSBuildThisFileDirectory)Extensions\AttributeExtensions.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)HostApp\Properties\PropertiesExtractor.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)HostApp\RhinoIdleManager.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)HostApp\RhinoLayerHelper.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)HostApp\RhinoObjectHelper.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)HostApp\RhinoUtils.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Mapper\Revit\LayerCategoryMapping.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Mapper\Revit\LayerOption.cs" />
